task-15-export-and-import: created modules for classes, enums and interfaces + interface DamageLogger exported as Logger from interfaces module
