+++ src/enums.ts
@@ -1,0 +1,9 @@
+enum Category {
+    JavaScript,
+    CSS,
+    HTML,
+    TypeScript,
+    Angular
+}
+
+export { Category };
